refactor: 카테고리 복수형 이름 변경
